Merging /w original repo.

+++ tests/gdx-tests/src/com/badlogic/gdx/tests/FilesTest.java
@@ -151,9 +151,11 @@ public class FilesTest extends GdxTest {
 		try {
 			testClasspath();
 			testInternal();
-			testExternal();
-			testAbsolute();
-			testLocal();
+			if (!(Gdx.app.getType() == ApplicationType.WebGL)) {
+				testExternal();
+				testAbsolute();
+				testLocal();
+			}
 		} catch (IOException ex) {
 			throw new RuntimeException(ex);
 		}
@@ -208,16 +210,13 @@ public class FilesTest extends GdxTest {
 			fail();
 		} catch (Exception ignored) {
 		}
-		FileHandle dir;
-		if (Gdx.app.getType() == ApplicationType.Android || Gdx.app.getType() == ApplicationType.iOS)
-			dir = Gdx.files.internal("data");
-		else
-			dir = Gdx.files.internal("../gdx-tests-android/assets/data");
+		FileHandle dir = Gdx.files.internal("data");
 		if (Gdx.app.getType() != ApplicationType.Android) {
 			if (!dir.exists()) fail();
 		}
 		if (!dir.isDirectory()) fail();
 		if (dir.list().length == 0) fail();
+		Gdx.app.log("FilesTest", "Files in data: " + Arrays.toString(dir.list()) + " (" + dir.list().length + ")");
 		FileHandle child = dir.child("badlogic.jpg");
 		if (!child.name().equals("badlogic.jpg")) fail();
 		if (!child.nameWithoutExtension().equals("badlogic")) fail();
@@ -225,14 +224,16 @@ public class FilesTest extends GdxTest {
 		if (Gdx.app.getType() != ApplicationType.Android) {
 			if (!child.parent().exists()) fail();
 		}
-		FileHandle copy = Gdx.files.external("badlogic.jpg-copy");
-		copy.delete();
-		if (copy.exists()) fail();
-		handle.copyTo(copy);
-		if (!copy.exists()) fail();
-		if (copy.length() != 68465) fail();
-		copy.delete();
-		if (copy.exists()) fail();
+		if (!(Gdx.app.getType() == ApplicationType.WebGL)) {
+			FileHandle copy = Gdx.files.external("badlogic.jpg-copy");
+			copy.delete();
+			if (copy.exists()) fail();
+			handle.copyTo(copy);
+			if (!copy.exists()) fail();
+			if (copy.length() != 68465) fail();
+			copy.delete();
+			if (copy.exists()) fail();
+		}
 		handle.read().close();
 		if (handle.readBytes().length != handle.length()) fail();
 	}
